Move isSupportedUserSearch from backend-core to shared-core.
